Improve c10 dispatcher lookup perf
Previously, looking up a kernel accidentally copied the DispatchTableEntry, which has as its member a std::function cache creator function.
Being an std::function, it was expensive to copy and cost us more than 50ns on each op call.
This diff fixes this by not copying DispatchTableEntry anymore.

Differential Revision: [D16910530](https://our.internmc.facebook.com/intern/diff/D16910530/)
